+++ GRDB/Migrations/Migration.swift
@@ -12,17 +12,37 @@
 
 /// An internal struct that defines a migration.
 struct Migration {
+    // PRAGMA foreign_key_check = ON was introduced in SQLite 3.7.16 http://www.sqlite.org/changes.html#version_3_7_16
+    // It is available from iOS 8.2 and OS X 10.10 https://github.com/yapstudios/YapDatabase/wiki/SQLite-version-(bundled-with-OS)
     let identifier: String
     let disabledForeignKeyChecks: Bool
     let migrate: (db: Database) throws -> Void
     
+    init(identifier: String, migrate: (db: Database) throws -> Void) {
+        self.identifier = identifier
+        self.disabledForeignKeyChecks = false
+        self.migrate = migrate
+    }
+    
+    @available(iOS 8.2, OSX 10.10, *)
+    init(identifier: String, disabledForeignKeyChecks: Bool, migrate: (db: Database) throws -> Void) {
+        self.identifier = identifier
+        self.disabledForeignKeyChecks = disabledForeignKeyChecks
+        self.migrate = migrate
+    }
+    
     func run(db: Database) throws {
-        if disabledForeignKeyChecks && Bool.fetchOne(db, "PRAGMA foreign_keys")! {
-            try runWithDisabledForeignKeys(db)
+        if #available(iOS 8.2, OSX 10.10, *) {
+            if disabledForeignKeyChecks && Bool.fetchOne(db, "PRAGMA foreign_keys")! {
+                try runWithDisabledForeignKeys(db)
+            } else {
+                try runWithoutDisabledForeignKeys(db)
+            }
         } else {
             try runWithoutDisabledForeignKeys(db)
         }
     }
+    
     
     private func runWithoutDisabledForeignKeys(db: Database) throws {
         try db.inTransaction(.Immediate) {
@@ -32,6 +52,7 @@ struct Migration {
         }
     }
     
+    @available(iOS 8.2, OSX 10.10, *)
     private func runWithDisabledForeignKeys(db: Database) throws {
         // Support for database alterations described at
         // https://www.sqlite.org/lang_altertable.html#otheralter